# apps/api/swagger.json
{"openapi":"3.0.0","paths":{"/v1/ai/nutritional-facts":{"post":{"operationId":"AiController_nutritionalFacts_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenerateNutritionalFactsDto"}}}}},"responses":{"200":{"description":"The generated nutritional facts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NutritionalFactsEntity"}}}}},"tags":["Ai"]}},"/v1/ai/tags":{"post":{"operationId":"AiController_tags_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateTagsDto"}}}},"responses":{"200":{"description":"The generated tags","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["Ai"]}},"/auth/google":{"get":{"operationId":"AuthController_googleLogin","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/google/callback":{"get":{"operationId":"AuthController_googleAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/status":{"get":{"operationId":"HealthCheckController_status","parameters":[],"responses":{"200":{"description":""}},"tags":["HealthCheck"]}},"/health":{"get":{"operationId":"HealthCheckController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["HealthCheck"]}},"/v1/recipes":{"get":{"operationId":"RecipesController_recipesList_v1","parameters":[],"responses":{"200":{"description":"The recipe records","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeListResponse"}}}}},"tags":["Recipes"]},"post":{"operationId":"RecipesController_createRecipe_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestRecipeEntity"}}}}},"tags":["Recipes"]}},"/v1/recipes/{id}":{"get":{"operationId":"RecipesController_recipe_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of recipe","schema":{"type":"string"}}],"responses":{"200":{"description":"The recipe record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeResponse"}}}}},"tags":["Recipes"]},"patch":{"operationId":"RecipesController_updateRecipe_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of recipe","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchRecipeDto"}}}},"responses":{"200":{"description":""}},"tags":["Recipes"]}},"/v1/tag-names":{"get":{"operationId":"TagsController_tagNameList_v1","parameters":[],"responses":{"200":{"description":"All tag names","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagNamesEntity"}}}}},"tags":["Tags"]}},"/v1/users/{id}":{"get":{"operationId":"UsersController_user_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of user","schema":{"type":"string"}}],"responses":{"200":{"description":"user's public info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublicResponse"}}}}},"tags":["Users"]}},"/v1/users/{id}/account":{"get":{"operationId":"UsersController_userAccount_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of user","schema":{"type":"string"}}],"responses":{"200":{"description":"user's account info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAccountResponse"}}}}},"tags":["Users"]},"patch":{"operationId":"UsersController_updateUser_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of user","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchUserDto"}}}},"responses":{"200":{"description":"user's info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAccountResponse"}}}}},"tags":["Users"]}},"/v1/users/{id}/followers":{"get":{"operationId":"UsersController_followers_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of user's followers","schema":{"type":"string"}}],"responses":{"200":{"description":"user's followers","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserFollowersResponse"}}}}}},"tags":["Users"]}},"/v1/users/{id}/follow":{"put":{"operationId":"UsersController_followUser_v1","parameters":[{"name":"id","required":true,"in":"path","description":"id of user to (un)follow","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchFollowUserDto"}}}},"responses":{"204":{"description":"(un)follow a user"}},"tags":["Users"]}}},"info":{"title":"Recipes API","description":"The recipes API description","version":"1.0","contact":{}},"tags":[{"name":"recipes","description":""}],"servers":[],"components":{"schemas":{"CreateIngredientDto":{"type":"object","properties":{"amount":{"type":"number"},"isFraction":{"type":"boolean"},"unit":{"type":"string","enum":["cups","fluidOunces","tablespoons","teaspoons","pints","quarts","gallons","pounds","ounces","liters","milliliters","kilograms","grams"],"nullable":true},"name":{"type":"string"}},"required":["amount","isFraction","unit","name"]},"GenerateNutritionalFactsDto":{"type":"object","properties":{"instruction":{"type":"string","nullable":true},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/CreateIngredientDto"}}},"required":["instruction","ingredients"]},"NutritionalFactsEntity":{"type":"object","properties":{"servings":{"type":"number","nullable":true},"servingAmount":{"type":"number","nullable":true},"servingUnit":{"type":"string","enum":["cups","fluidOunces","tablespoons","teaspoons","pints","quarts","gallons","pounds","ounces","liters","milliliters","kilograms","grams"],"nullable":true},"proteinInG":{"type":"number","nullable":true},"totalFatInG":{"type":"number","nullable":true},"carbohydratesInG":{"type":"number","nullable":true},"fiberInG":{"type":"number","nullable":true},"sugarInG":{"type":"number","nullable":true},"sodiumInMg":{"type":"number","nullable":true},"cholesterolInMg":{"type":"number","nullable":true},"saturatedFatInG":{"type":"number","nullable":true},"transFatInG":{"type":"number","nullable":true},"potassiumInMg":{"type":"number","nullable":true},"vitaminAInIU":{"type":"number","nullable":true},"vitaminCInMg":{"type":"number","nullable":true},"calciumInMg":{"type":"number","nullable":true},"ironInMg":{"type":"number","nullable":true},"vitaminDInIU":{"type":"number","nullable":true},"vitaminB6InMg":{"type":"number","nullable":true},"vitaminB12InMg":{"type":"number","nullable":true},"magnesiumInMg":{"type":"number","nullable":true},"folateInMcg":{"type":"number","nullable":true},"thiaminInMg":{"type":"number","nullable":true},"riboflavinInMg":{"type":"number","nullable":true},"niacinInMg":{"type":"number","nullable":true},"caloriesInKcal":{"type":"number","nullable":true}},"required":["servings","servingAmount","servingUnit","proteinInG","totalFatInG","carbohydratesInG","fiberInG","sugarInG","sodiumInMg","cholesterolInMg","saturatedFatInG","transFatInG","potassiumInMg","vitaminAInIU","vitaminCInMg","calciumInMg","ironInMg","vitaminDInIU","vitaminB6InMg","vitaminB12InMg","magnesiumInMg","folateInMcg","thiaminInMg","riboflavinInMg","niacinInMg","caloriesInKcal"]},"GenerateBaseDto":{"type":"object","properties":{"instruction":{"type":"string","nullable":true},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/CreateIngredientDto"}}},"required":["instruction","ingredients"]},"GenerateTagsDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string","nullable":true},"steps":{"type":"array","items":{"$ref":"#/components/schemas/GenerateBaseDto"}}},"required":["name","description","steps"]},"RecipeUserEntity":{"type":"object","properties":{"id":{"type":"string"},"handle":{"type":"string"},"imageUrl":{"type":"string","nullable":true}},"required":["id","handle","imageUrl"]},"RecipeMinimalResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"preparationTimeInMinutes":{"type":"integer","nullable":true},"cookingTimeInMinutes":{"type":"integer","nullable":true},"imageUrl":{"type":"string","nullable":true},"tags":{"type":"array","items":{"type":"string"}},"user":{"$ref":"#/components/schemas/RecipeUserEntity"}},"required":["id","name","description","preparationTimeInMinutes","cookingTimeInMinutes","imageUrl","tags","user"]},"PaginationResponse":{"type":"object","properties":{"totalRecords":{"type":"number"},"currentCursor":{"type":"string","nullable":true},"nextCursor":{"type":"string","nullable":true}},"required":["totalRecords","currentCursor","nextCursor"]},"RecipeListResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RecipeMinimalResponse"}},"pagination":{"$ref":"#/components/schemas/PaginationResponse"}},"required":["data","pagination"]},"IngredientEntity":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"amount":{"type":"number"},"isFraction":{"type":"boolean"},"unit":{"type":"string","enum":["cups","fluidOunces","tablespoons","teaspoons","pints","quarts","gallons","pounds","ounces","liters","milliliters","kilograms","grams"],"nullable":true},"name":{"type":"string"}},"required":["id","createdAt","updatedAt","amount","isFraction","unit","name"]},"StepEntity":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"instruction":{"type":"string","nullable":true},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/IngredientEntity"}},"imageUrl":{"type":"string","nullable":true}},"required":["id","createdAt","updatedAt","instruction","ingredients","imageUrl"]},"RecipeResponse":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"preparationTimeInMinutes":{"type":"integer","nullable":true},"cookingTimeInMinutes":{"type":"integer","nullable":true},"imageUrl":{"type":"string","nullable":true},"equipments":{"type":"array","items":{"type":"string"}},"steps":{"type":"array","items":{"$ref":"#/components/schemas/StepEntity"}},"nutritionalFacts":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/NutritionalFactsEntity"}]},"tags":{"type":"array","items":{"type":"string"}},"user":{"$ref":"#/components/schemas/RecipeUserEntity"},"isPublic":{"type":"boolean"}},"required":["id","createdAt","updatedAt","name","description","preparationTimeInMinutes","cookingTimeInMinutes","imageUrl","equipments","steps","nutritionalFacts","tags","user","isPublic"]},"CreateStepDto":{"type":"object","properties":{"instruction":{"type":"string","nullable":true},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/CreateIngredientDto"}},"base64Image":{"type":"string","nullable":true}},"required":["instruction","ingredients","base64Image"]},"NutritionalFactsDto":{"type":"object","properties":{"servings":{"type":"number","nullable":true},"servingAmount":{"type":"number","nullable":true},"servingUnit":{"type":"string","enum":["cups","fluidOunces","tablespoons","teaspoons","pints","quarts","gallons","pounds","ounces","liters","milliliters","kilograms","grams"],"nullable":true},"proteinInG":{"type":"number","nullable":true},"totalFatInG":{"type":"number","nullable":true},"carbohydratesInG":{"type":"number","nullable":true},"fiberInG":{"type":"number","nullable":true},"sugarInG":{"type":"number","nullable":true},"sodiumInMg":{"type":"number","nullable":true},"cholesterolInMg":{"type":"number","nullable":true},"saturatedFatInG":{"type":"number","nullable":true},"transFatInG":{"type":"number","nullable":true},"potassiumInMg":{"type":"number","nullable":true},"vitaminAInIU":{"type":"number","nullable":true},"vitaminCInMg":{"type":"number","nullable":true},"calciumInMg":{"type":"number","nullable":true},"ironInMg":{"type":"number","nullable":true},"vitaminDInIU":{"type":"number","nullable":true},"vitaminB6InMg":{"type":"number","nullable":true},"vitaminB12InMg":{"type":"number","nullable":true},"magnesiumInMg":{"type":"number","nullable":true},"folateInMcg":{"type":"number","nullable":true},"thiaminInMg":{"type":"number","nullable":true},"riboflavinInMg":{"type":"number","nullable":true},"niacinInMg":{"type":"number","nullable":true},"caloriesInKcal":{"type":"number","nullable":true}},"required":["servings","servingAmount","servingUnit","proteinInG","totalFatInG","carbohydratesInG","fiberInG","sugarInG","sodiumInMg","cholesterolInMg","saturatedFatInG","transFatInG","potassiumInMg","vitaminAInIU","vitaminCInMg","calciumInMg","ironInMg","vitaminDInIU","vitaminB6InMg","vitaminB12InMg","magnesiumInMg","folateInMcg","thiaminInMg","riboflavinInMg","niacinInMg","caloriesInKcal"]},"CreateRecipeDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string","nullable":true},"base64Image":{"type":"string","nullable":true},"preparationTimeInMinutes":{"type":"number","nullable":true},"cookingTimeInMinutes":{"type":"number","nullable":true},"steps":{"type":"array","items":{"$ref":"#/components/schemas/CreateStepDto"}},"nutritionalFacts":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/NutritionalFactsDto"}]},"tags":{"type":"array","items":{"type":"string"}},"equipments":{"type":"array","items":{"type":"string"}},"isPublic":{"type":"boolean"}},"required":["name","description","base64Image","preparationTimeInMinutes","cookingTimeInMinutes","steps","nutritionalFacts","tags","equipments","isPublic"]},"BadRequestIngredientEntity":{"type":"object","properties":{"amount":{"type":"string"},"unit":{"type":"string"},"name":{"type":"string"}}},"BadRequestStepEntity":{"type":"object","properties":{"instruction":{"type":"string"},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/BadRequestIngredientEntity"}},"base64Image":{"type":"string"}},"required":["base64Image"]},"BadRequestNutritionalFactsEntity":{"type":"object","properties":{"servings":{"type":"string"},"servingAmount":{"type":"string"},"servingUnit":{"type":"string"},"proteinInG":{"type":"string"},"totalFatInG":{"type":"string"},"carbohydratesInG":{"type":"string"},"fiberInG":{"type":"string"},"sugarInG":{"type":"string"},"sodiumInMg":{"type":"string"},"cholesterolInMg":{"type":"string"},"saturatedFatInG":{"type":"string"},"transFatInG":{"type":"string"},"potassiumInMg":{"type":"string"},"vitaminAInIU":{"type":"string"},"vitaminCInMg":{"type":"string"},"calciumInMg":{"type":"string"},"ironInMg":{"type":"string"},"vitaminDInIU":{"type":"string"},"vitaminB6InMg":{"type":"string"},"vitaminB12InMg":{"type":"string"},"magnesiumInMg":{"type":"string"},"folateInMcg":{"type":"string"},"thiaminInMg":{"type":"string"},"riboflavinInMg":{"type":"string"},"niacinInMg":{"type":"string"},"caloriesInKcal":{"type":"string"},"fatInG":{"type":"string"},"fiber":{"type":"string"},"sugar":{"type":"string"}}},"BadRequestRecipeEntity":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"base64Image":{"type":"string"},"preparationTimeInMinutes":{"type":"string"},"cookingTimeInMinutes":{"type":"string"},"steps":{"type":"array","items":{"$ref":"#/components/schemas/BadRequestStepEntity"}},"nutritionalFacts":{"$ref":"#/components/schemas/BadRequestNutritionalFactsEntity"},"tags":{"type":"array","items":{"type":"string"}},"equipments":{"type":"array","items":{"type":"string"}},"isPublic":{"type":"string"}}},"PatchIngredientDto":{"type":"object","properties":{"id":{"type":"string"},"amount":{"type":"number"},"isFraction":{"type":"boolean"},"unit":{"type":"string","enum":["cups","fluidOunces","tablespoons","teaspoons","pints","quarts","gallons","pounds","ounces","liters","milliliters","kilograms","grams"],"nullable":true},"name":{"type":"string"}},"required":["id","amount","isFraction","unit","name"]},"PatchStepDto":{"type":"object","properties":{"id":{"type":"string"},"instruction":{"type":"string"},"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/PatchIngredientDto"}},"base64Image":{"type":"string","nullable":true}},"required":["ingredients","base64Image"]},"PatchRecipeDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string","nullable":true},"base64Image":{"type":"string"},"preparationTimeInMinutes":{"type":"number","nullable":true},"cookingTimeInMinutes":{"type":"number","nullable":true},"steps":{"type":"array","items":{"$ref":"#/components/schemas/PatchStepDto"}},"nutritionalFacts":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/NutritionalFactsDto"}]},"tags":{"type":"array","items":{"type":"string"}},"equipments":{"type":"array","items":{"type":"string"}},"isPublic":{"type":"boolean"}},"required":["base64Image"]},"TagNamesEntity":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string"}},"pagination":{"$ref":"#/components/schemas/PaginationResponse"}},"required":["data","pagination"]},"UserPublicResponse":{"type":"object","properties":{"handle":{"type":"string"},"imageUrl":{"type":"string","nullable":true},"followers":{"type":"number"},"followings":{"type":"number"},"recipes":{"type":"number"},"amIFollowing":{"type":"boolean"}},"required":["handle","imageUrl","followers","followings","recipes"]},"UiTheme":{"type":"string","enum":["system","light","dark"]},"NumberFormat":{"type":"string","enum":["default","decimal","fraction"]},"MeasurementFormat":{"type":"string","enum":["default","imperial","metric"]},"UserAccountResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"handle":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"diet":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/NutritionalFactsEntity"}]},"uiTheme":{"allOf":[{"$ref":"#/components/schemas/UiTheme"}]},"numberFormat":{"allOf":[{"$ref":"#/components/schemas/NumberFormat"}]},"measurementFormat":{"allOf":[{"$ref":"#/components/schemas/MeasurementFormat"}]},"imageUrl":{"type":"string","nullable":true}},"required":["id","name","email","handle","createdAt","updatedAt","diet","uiTheme","numberFormat","measurementFormat","imageUrl"]},"PatchUserDto":{"type":"object","properties":{"name":{"type":"string"},"handle":{"type":"string"},"uiTheme":{"allOf":[{"$ref":"#/components/schemas/UiTheme"}]},"numberFormat":{"allOf":[{"$ref":"#/components/schemas/NumberFormat"}]},"measurementFormat":{"allOf":[{"$ref":"#/components/schemas/MeasurementFormat"}]}}},"UserFollowersResponse":{"type":"object","properties":{"id":{"type":"string"},"handle":{"type":"string"},"imageUrl":{"type":"string","nullable":true}},"required":["id","handle","imageUrl"]},"PatchFollowUserDto":{"type":"object","properties":{"follow":{"type":"boolean"}},"required":["follow"]}}}}